requirements.txt
contourpy==1.1.0
cycler==0.11.0
fonttools==4.41.0
grpcio==1.56.0
grpcio-tools==1.56.0
importlib-resources==6.0.0
joblib==1.3.1
kiwisolver==1.4.4
matplotlib==3.7.2
numpy
packaging==23.1
Pillow==10.0.0
protobuf==4.23.4
psycopg2-binary==2.9.6
pyparsing==3.0.9
python-dateutil==2.8.2
scikit-learn==1.3.0
scipy
six==1.16.0
scikit-learn
threadpoolctl==3.2.0
zipp==3.16.2
